Move `hash2field` into `hash2curve`
