+++ elliptic-curve/src/hash2curve.rs
@@ -3,12 +3,13 @@
 //! <https://datatracker.ietf.org/doc/draft-irtf-cfrg-hash-to-curve>
 
 mod group_digest;
+mod hash2field;
 mod isogeny;
 mod map2curve;
 mod osswu;
 
-pub use crate::hash2field::*;
 pub use group_digest::*;
+pub use hash2field::*;
 pub use isogeny::*;
 pub use map2curve::*;
 pub use osswu::*;
